Refactor CodeProcessor into CodeAgent with pipeline stages.
