+++ src/jrdev/commands/code.py
@@ -7,7 +7,7 @@ Code command implementation for the JrDev application.
 from typing import Any, List
 
 from jrdev.ui.ui import PrintType
-from jrdev.services.code_processor import CodeProcessor
+from jrdev.agents.code_agent import CodeAgent
 from jrdev.core.exceptions import CodeTaskCancelled
 
 
@@ -16,7 +16,7 @@ async def handle_code(app: Any, args: List[str], worker_id: str) -> None:
         app.ui.print_text("Usage: /code <message>", print_type=PrintType.ERROR)
         return
     message = " ".join(args[1:])
-    code_processor = CodeProcessor(app, worker_id)
+    code_processor = CodeAgent(app, worker_id)
     try:
         await code_processor.process(message)
     except CodeTaskCancelled:
